Adds the 'push_to_discourse' flag to users on membership sync.
See https://github.com/United-Philanthropy-Forum/km-collaborative/issues/1025 for background.
This addresses an issue where, after joining or being assigned a group, a Drupal user's
account gets pushed to Discourse, but isn't flagged to be synced. On subsequent sync
actions, the module tries to create a new account because of the missing flag, which fails
because the user already exists.